made session being created only on message send

+++ pink_agent.py
@@ -12,9 +12,8 @@ class PinkAgent:
     def __init__(self):
         self.logger = logging.Logger(__name__)
         
-        session_id = str(uuid.uuid4())
-        self.session_filepath = os.path.join(get_tempdir(),session_id + ".db")
-        self.session = SQLiteSession(session_id,self.session_filepath)
+        self.session = None
+        self.session_filepath = None
         
         self.mcp_servers = []
         self.mcp_agent = None
@@ -25,6 +24,12 @@ class PinkAgent:
             instructions="You are an assistance, if the user will address any questions for which MCP Server access is require you will handoff to tools_agent. The tools you have access include Time Server, to answer time related queries",
             handoffs=[],
         )
+    # create session only when chat is sent
+    async def init_session(self):
+        session_id = str(uuid.uuid4())
+        self.session_filepath = os.path.join(get_tempdir(),session_id + ".db")
+        self.session = SQLiteSession(session_id,self.session_filepath)
+        
     async def load_mcp_servers(self, servers):
         if not servers:
             return
@@ -71,12 +76,15 @@ class PinkAgent:
         print("[*]Reloaded MCP Servers")
         
     async def chat(self,input):
+        if not self.session:
+            await self.init_session()
         result = await Runner.run(self.triage_agent,input=input,session=self.session)
         # need to add error handling
         return result.final_output
     
     async def cleanup(self):
-        self.session.close()
+        if self.session:
+            self.session.close()
         if self.mcp_servers:
             for server in self.mcp_servers:
                 try:
@@ -85,8 +93,9 @@ class PinkAgent:
                     self.logger.info("[*]MCP cleanup interrupted (normal during shutdown)")
                 except Exception as e:
                     self.logger.error(f"[*]MCP cleanup error: {e}")
-        time.sleep(0.2)
-        os.unlink(self.session_filepath)
+        if self.session_filepath:
+            time.sleep(0.2)
+            os.unlink(self.session_filepath)
        
     def get_mcp_servers(self):
            if self.mcp_servers:
